modules/mdm/src/test/java/com/ibm/cloud/mdm/v0/model/GetSummaryTest.java
/*
 * (C) Copyright IBM Corp. 2021.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations under the License.
 */

package com.ibm.cloud.mdm.v0.model;

import com.ibm.cloud.mdm.v0.model.GetSummary;
import com.ibm.cloud.mdm.v0.model.GetSummaryEntities;
import com.ibm.cloud.mdm.v0.model.GetSummaryEntitySizeStats;
import com.ibm.cloud.mdm.v0.model.GetSummaryStatsAndEntities;
import com.ibm.cloud.mdm.v0.utils.TestUtilities;
import com.ibm.cloud.sdk.core.service.model.FileWithMetadata;
import java.io.InputStream;
import java.util.ArrayList;
import java.util.Arrays;
import java.util.HashMap;
import java.util.List;
import org.testng.annotations.Test;
import static org.testng.Assert.*;

/**
 * Unit test class for the GetSummary model.
 */
public class GetSummaryTest {
  final HashMap<String, InputStream> mockStreamMap = TestUtilities.createMockStreamMap();
  final List<FileWithMetadata> mockListFileWithMetadata = TestUtilities.creatMockListFileWithMetadata();

  @Test
  public void testGetSummary() throws Throwable {
    GetSummary getSummaryModel = new GetSummary();
    assertNull(getSummaryModel.getTotalRecords());
    assertNull(getSummaryModel.getSingletonEntityCount());
    assertNull(getSummaryModel.getStatsAndEntities());
    assertNull(getSummaryModel.getUniqueSourceCount());
  }
}